WIP merge mblockalloc into memory
