+++ asterius/rts/rts.memory.mjs
@@ -1,11 +1,17 @@
 import * as rtsConstants from "./rts.constants.mjs";
 
+function mask(n) {
+  return (BigInt(1) << BigInt(n)) - BigInt(1);
+}
+
 export class Memory {
   constructor() {
     this.memory = undefined;
     this.staticMBlocks = undefined;
     this.i8View = undefined;
     this.dataView = undefined;
+    this.capacity = undefined;
+    this.bitset = undefined;
     Object.seal(this);
   }
 
@@ -13,6 +19,8 @@ export class Memory {
     this.memory = memory;
     this.staticMBlocks = static_mblocks;
     this.initView();
+    this.capacity = this.buffer.byteLength / rtsConstants.mblock_size;
+    this.bitset = mask(this.capacity);
   }
 
   initView() {
@@ -165,6 +173,29 @@ export class Memory {
       Memory.unTag(p) >=
       this.staticMBlocks << Math.log2(rtsConstants.mblock_size)
     );
+  }
+
+  getMBlocks(n) {
+    const m = mask(n);
+    for (let i = BigInt(0); i <= BigInt(this.capacity - n); ++i) {
+      const mi = m << i;
+      if (!(this.bitset & mi)) {
+        this.bitset |= mi;
+        return Memory.tagData(Number(i) * rtsConstants.mblock_size);
+      }
+    }
+    const d = Math.max(n, this.capacity),
+      prev_capacity = this.capacity;
+    this.grow(d * (rtsConstants.mblock_size / rtsConstants.pageSize));
+    this.capacity += d;
+    this.bitset |= m << BigInt(prev_capacity);
+    return Memory.tagData(prev_capacity * rtsConstants.mblock_size);
+  }
+
+  freeMBlocks(p, n) {
+    const mblock_no =
+      BigInt(Memory.unTag(p)) >> BigInt(Math.log2(rtsConstants.mblock_size));
+    this.bitset &= ~(mask(n) << mblock_no);
   }
 
   strlen(_str) {
